+++ src/main/java/co/elastic/logstash/filters/elasticintegration/geoip/GeoipDatabaseHolder.java
@@ -1,0 +1,10 @@
+package co.elastic.logstash.filters.elasticintegration.geoip;
+
+import org.elasticsearch.ingest.geoip.GeoIpDatabase;
+
+interface GeoipDatabaseHolder {
+    boolean isValid();
+    GeoIpDatabase getDatabase();
+    String getTypeIdentifier();
+    String info();
+}